Correctly report groups for FreeIPAPlugin
dn is not a searchable field, as far as I can tell, with searches by dn
always yielding no results on my FreeIPA installation.

Instead, split the DNs into a base DN and an RDN, and build a search for
each listed base DN to return all the groups under that, based on the
RDNs.